+++ hypervisor/debug/Makefile
@@ -4,6 +4,7 @@
 
 FILE_PATH := $(dir $(MKFL_NAME))
 SRCS += $(wildcard $(FILE_PATH)/*.c)
+SRCS += $(wildcard $(FILE_PATH)/$(ARCH)/*.c)
 OBJS += $(patsubst %.c,$(HV_OBJDIR)/%.o,$(SRCS))
 
 .PHONY: default
