+++ hypervisor/debug/riscv/npk_log.c
