+++ hypervisor/debug/x86/dump.c
