+++ hypervisor/debug/x86/npk_log.c
