+++ hypervisor/debug/x86/profiling.c
